backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
asyncpg==0.30.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
mem0ai==1.0.11
psycopg[pool]==3.2.4
openai>=1.90.0
bcrypt==4.3.0
PyJWT==2.10.1
